Move `addIssues.jsx` into the `addIssues` folder
There were several new files created related to `addIssues` and it
felt like time to make a new directory to house these files.

One thing that could be done in the future is to use more absolute
paths instead of relative paths so that in the future when files and
folders get moved around we don't have to edit file paths.
